dhclient: Fix the hook
Because the hook is sourced `.` by dhclient-script, not executed as a
child having a statement like `exit 0` in there causes dhclient-script
to stop executing after that. This means that since 6rd is likely one of
the first hooks to fire any further hooks which are asciibetically
sorted later than 6rd won't be triggered. This leads to things like ntp
servers not getting updated and other stuff. It's nasty.

So now what we have instead is one big hairy ball of `if/elif/else`
statemens with `return` in them instead.

Since it's reasonable to assume people will have sixrd and grep
installed on their systems I've also removed the prereq checks.

Also fixes one shellcheck error.